Refactor code into api from internal
