+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/rendezvous/model/EmbeddedRendezvous.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Matrix.org Foundation C.I.C.
+ * Copyright 2022 The Matrix.org Foundation C.I.C.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package org.matrix.android.sdk.internal.rendezvous.model
+package org.matrix.android.sdk.api.rendezvous.model
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 open class EmbeddedRendezvous(
-    @Json(name = "transport") val transport: RendezvousTransportDetails,
-    @Json(name = "algorithm") val algorithm: SecureRendezvousChannelAlgorithm
+        @Json(name = "transport") val transport: RendezvousTransportDetails,
+        @Json(name = "algorithm") val algorithm: SecureRendezvousChannelAlgorithm
 )
