+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/api/rendezvous/model/RendezvousError.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Matrix.org Foundation C.I.C.
+ * Copyright 2022 The Matrix.org Foundation C.I.C.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,9 @@
  * limitations under the License.
  */
 
-package org.matrix.android.sdk.internal.rendezvous
+package org.matrix.android.sdk.api.rendezvous.model
 
-import okhttp3.MediaType
-import org.matrix.android.sdk.internal.rendezvous.model.RendezvousTransportDetails
+import org.matrix.android.sdk.api.rendezvous.RendezvousFailureReason
 
-interface RendezvousTransport {
-    var ready: Boolean;
-    var onCancelled: ((reason: RendezvousFailureReason) -> Unit)?;
-    suspend fun details(): RendezvousTransportDetails;
-    suspend fun send(contentType: MediaType, data: ByteArray);
-    suspend fun receive(): ByteArray?;
-    suspend fun cancel(reason: RendezvousFailureReason);
+class RendezvousError(val description: String, val reason: RendezvousFailureReason): RuntimeException(description) {
 }
